product_review_search/requirements.txt
streamlit==1.35.0
pandas==2.2.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pyarrow==16.1.0